pulling docs out into a seperate file to clean up cli
