+++ src/cmd/singularity/cli/actions.go
@@ -17,107 +17,10 @@ import (
 	"github.com/singularityware/singularity/src/pkg/buildcfg"
 	runtimeconfig "github.com/singularityware/singularity/src/runtime/workflows/workflows/singularity/config"
 
-	"github.com/singularityware/singularity/docs"
 	"github.com/spf13/cobra"
+
+	"github.com/singularityware/singularity/docs"
 )
-
-var formats string = `
-
-*.sqsh              SquashFS format.  Native to Singularity 2.4+
-
-*.img               This is the native Singularity image format for all
-                    Singularity versions < 2.4.
-
-*.tar\*              Tar archives are exploded to a temporary directory and
-                    run within that directory (and cleaned up after). The
-                    contents of the archive is a root file system with root
-                    being in the current directory. All compression
-                    suffixes are supported.
-
-directory/          Container directories that contain a valid root file
-                    system.
-
-instance://*        A local running instance of a container. (See the
-                    instance command group.)
-
-shub://*            A container hosted on Singularity Hub
-
-docker://*          A container hosted on Docker Hub`
-
-var execUse string = `exec [exec options...] <container> ...`
-
-var execShort string = `Execute a command within container`
-
-var execLong string = `
-singularity exec supports the following formats:` + formats
-
-var execExamples string = `
-$ singularity exec /tmp/Debian.img cat /etc/debian_version
-$ singularity exec /tmp/Debian.img python ./hello_world.py
-$ cat hello_world.py | singularity exec /tmp/Debian.img python
-$ sudo singularity exec --writable /tmp/Debian.img apt-get update
-$ singularity exec instance://my_instance ps -ef`
-
-var shellUse string = `shell [shell options...] <container>`
-
-var shellShort string = `Run a Bourne shell within container`
-
-var shellLong string = `
-singularity shell supports the following formats:` + formats
-
-var shellExamples string = `
-$ singularity shell /tmp/Debian.img
-Singularity/Debian.img> pwd
-/home/gmk/test
-Singularity/Debian.img> exit
-
-$ singularity shell -C /tmp/Debian.img
-Singularity/Debian.img> pwd
-/home/gmk
-Singularity/Debian.img> ls -l
-total 0
-Singularity/Debian.img> exit
-
-$ sudo singularity shell -w /tmp/Debian.img
-$ sudo singularity shell --writable /tmp/Debian.img
-
-$ singularity shell instance://my_instance 
-
-$ singularity shell instance://my_instance
-Singularity: Invoking an interactive shell within container...
-Singularity container:~> ps -ef
-UID        PID  PPID  C STIME TTY          TIME CMD
-ubuntu       1     0  0 20:00 ?        00:00:00 /usr/local/bin/singularity/bin/sinit
-ubuntu       2     0  0 20:01 pts/8    00:00:00 /bin/bash --norc
-ubuntu       3     2  0 20:02 pts/8    00:00:00 ps -ef`
-
-var runUse string = `run [run options...] <container>`
-
-var runShort string = `Launch a runscript within container`
-
-var runLong string = `
-This command will launch a Singularity container and execute a runscript
-if one is defined for that container. The runscript is a metadata file within
-the container that containes shell commands. If the file is present (and 
-executable) then this command will execute that file within the container 
-automatically. All arguments following the container name will be passed 
-directly to the runscript.
-
-singularity run accepts the following container formats:` + formats
-
-var runExamples string = `
-# Here we see that the runscript prints "Hello world: "
-$ singularity exec /tmp/Debian.img cat /singularity
-#!/bin/sh
-echo "Hello world: "
-
-# It runs with our inputs when we run the image
-$ singularity run /tmp/Debian.img one two three
-Hello world: one two three
-
-# Note that this does the same thing
-$ ./tmp/Debian.img one two three
-`
 
 func init() {
 	actionCmds := []*cobra.Command{
@@ -128,21 +31,6 @@ func init() {
 
 	// TODO : the next n lines of code are repeating too much but I don't
 	// know how to shorten them tonight
-	execHelp := func(c *cobra.Command, args []string) {
-		docs.DispManPg("singularity-exec")
-	}
-	ExecCmd.SetHelpFunc(execHelp)
-
-	shellHelp := func(c *cobra.Command, args []string) {
-		docs.DispManPg("singularity-shell")
-	}
-	ShellCmd.SetHelpFunc(shellHelp)
-
-	runHelp := func(c *cobra.Command, args []string) {
-		docs.DispManPg("singularity-run")
-	}
-	RunCmd.SetHelpFunc(runHelp)
-
 	for _, cmd := range actionCmds {
 		cmd.PersistentFlags().AddFlag(actionFlags.Lookup("bind"))
 		cmd.PersistentFlags().AddFlag(actionFlags.Lookup("contain"))
@@ -183,10 +71,10 @@ var ExecCmd = &cobra.Command{
 		execWrapper(cmd, args[0], a)
 	},
 
-	Use:     execUse,
-	Short:   execShort,
-	Long:    execLong,
-	Example: execExamples,
+	Use:     docs.ExecUse,
+	Short:   docs.ExecShort,
+	Long:    docs.ExecLong,
+	Example: docs.ExecExamples,
 }
 
 // shellCmd represents the shell command
@@ -198,10 +86,10 @@ var ShellCmd = &cobra.Command{
 		execWrapper(cmd, args[0], a)
 	},
 
-	Use:     shellUse,
-	Short:   shellShort,
-	Long:    shellLong,
-	Example: shellExamples,
+	Use:     docs.ShellUse,
+	Short:   docs.ShellShort,
+	Long:    docs.ShellLong,
+	Example: docs.ShellExamples,
 }
 
 // runCmd represents the run command
@@ -213,10 +101,10 @@ var RunCmd = &cobra.Command{
 		execWrapper(cmd, args[0], a)
 	},
 
-	Use:     runUse,
-	Short:   runShort,
-	Long:    runLong,
-	Example: runExamples,
+	Use:     docs.RunUse,
+	Short:   docs.RunShort,
+	Long:    docs.RunLong,
+	Example: docs.RunExamples,
 }
 
 // TODO: Let's stick this in another file so that that CLI is just CLI
