+++ src/cmd/singularity/cli/capability.go
@@ -9,33 +9,13 @@
 package cli
 
 import (
-	"github.com/singularityware/singularity/docs"
 	"github.com/spf13/cobra"
+
+	"github.com/singularityware/singularity/docs"
 )
 
-var capabilityUse string = `capability <subcommand>`
-
-var capabilityShort string = `Manage Linux capabilities`
-
-var capabilityLong string = `
-Capabilities allow you to have fine grained control over the permissions that 
-your containers need to run. For instance, if you need to `
-
-var capabilityExample string = `
-All group commands have their own help output:
-
-$ singularity help capability add
-$ singularity capability list --help
-`
-
 func init() {
-
-	manHelp := func(c *cobra.Command, args []string) {
-		docs.DispManPg("singularity-capability")
-	}
-
 	SingularityCmd.AddCommand(CapabilityCmd)
-	CapabilityCmd.SetHelpFunc(manHelp)
 	CapabilityCmd.AddCommand(CapabilityAddCmd)
 	CapabilityCmd.AddCommand(CapabilityDropCmd)
 	CapabilityCmd.AddCommand(CapabilityListCmd)
@@ -45,8 +25,8 @@ var CapabilityCmd = &cobra.Command{
 	Run: nil,
 	DisableFlagsInUseLine: true,
 
-	Use:     capabilityUse,
-	Short:   capabilityShort,
-	Long:    capabilityLong,
-	Example: capabilityExample,
+	Use:     docs.CapabilityUse,
+	Short:   docs.CapabilityShort,
+	Long:    docs.CapabilityLong,
+	Example: docs.CapabilityExample,
 }
