+++ src/cmd/singularity/cli/capability_drop.go
@@ -10,43 +10,12 @@ package cli
 import (
 	"fmt"
 
-	"github.com/singularityware/singularity/docs"
 	"github.com/spf13/cobra"
+
+	"github.com/singularityware/singularity/docs"
 )
 
-var capabilityDropUse string = `drop [drop options...] <capabilities>`
-
-var capabilityDropShort string = `
-remove Linux capabilities from your container at runtime`
-
-var capabilityDropLong string = `
-The capability drop command allows you to remove Linux capabilities from your 
-container with fine grained precision. This way you can ensure that your 
-container is as secure as it can be given the functions it must carry out. For 
-instance, `
-
-var capabilityDropExample string = `
-$ singularity capability.drop /tmp/my-sql.img mysql
-
-$ singularity shell capability://mysql
-Singularity my-sql.img> pwd
-/home/mibauer/mysql
-Singularity my-sql.img> ps
-PID TTY          TIME CMD
-1 pts/0    00:00:00 sinit
-2 pts/0    00:00:00 bash
-3 pts/0    00:00:00 ps
-Singularity my-sql.img> 
-
-$ singularity capability.stop /tmp/my-sql.img mysql
-Stopping /tmp/my-sql.img mysql`
-
 func init() {
-
-	manHelp := func(c *cobra.Command, args []string) {
-		docs.DispManPg("singularity-capability-add")
-	}
-
 	capabilityDropCmds := []*cobra.Command{
 		CapabilityDropCmd,
 		// capabilityDotDropCmd,
@@ -54,7 +23,6 @@ func init() {
 
 	for _, cmd := range capabilityDropCmds {
 		cmd.Flags().SetInterspersed(false)
-		cmd.SetHelpFunc(manHelp)
 	}
 
 	// SingularityCmd.AddCommand(capabilityDotDropCmd)
@@ -67,10 +35,10 @@ var CapabilityDropCmd = &cobra.Command{
 		fmt.Println("dropping capability")
 	},
 
-	Use:     capabilityDropUse,
-	Short:   capabilityDropShort,
-	Long:    capabilityDropLong,
-	Example: capabilityDropExample,
+	Use:     docs.CapabilityDropUse,
+	Short:   docs.CapabilityDropShort,
+	Long:    docs.CapabilityDropLong,
+	Example: docs.CapabilityDropExample,
 }
 
 /*
