+++ src/cmd/singularity/cli/capability_list.go
@@ -10,40 +10,12 @@ package cli
 import (
 	"fmt"
 
-	"github.com/singularityware/singularity/docs"
 	"github.com/spf13/cobra"
+
+	"github.com/singularityware/singularity/docs"
 )
 
-var capabilityListUse string = `list [list options...] <capabilities>`
-
-var capabilityListShort string = `
-list Linux capabilities on a container`
-
-var capabilityListLong string = `The capability list command allows you to see 
-what Linux capabilities are associated with your container.`
-
-var capabilityListExample string = `
-$ singularity capability.list /tmp/my-sql.img mysql
-
-$ singularity shell capability://mysql
-Singularity my-sql.img> pwd
-/home/mibauer/mysql
-Singularity my-sql.img> ps
-PID TTY          TIME CMD
-  1 pts/0    00:00:00 sinit
-  2 pts/0    00:00:00 bash
-  3 pts/0    00:00:00 ps
-Singularity my-sql.img> 
-
-$ singularity capability.stop /tmp/my-sql.img mysql
-Stopping /tmp/my-sql.img mysql`
-
 func init() {
-
-	manHelp := func(c *cobra.Command, args []string) {
-		docs.DispManPg("singularity-capability-list")
-	}
-
 	capabilityListCmds := []*cobra.Command{
 		CapabilityListCmd,
 		// capabilityDotListCmd,
@@ -51,7 +23,6 @@ func init() {
 
 	for _, cmd := range capabilityListCmds {
 		cmd.Flags().SetInterspersed(false)
-		cmd.SetHelpFunc(manHelp)
 	}
 
 	// SingularityCmd.AddCommand(capabilityDotListCmd)
@@ -64,10 +35,10 @@ var CapabilityListCmd = &cobra.Command{
 		fmt.Println("listping capability")
 	},
 
-	Use:     capabilityListUse,
-	Short:   capabilityListShort,
-	Long:    capabilityListLong,
-	Example: capabilityListExample,
+	Use:     docs.CapabilityListUse,
+	Short:   docs.CapabilityListShort,
+	Long:    docs.CapabilityListLong,
+	Example: docs.CapabilityListExample,
 }
 
 /*
