+++ src/cmd/singularity/cli/pull.go
@@ -11,7 +11,7 @@ import (
 	"github.com/singularityware/singularity/src/pkg/libexec"
 	"github.com/spf13/cobra"
 
-	"github.com/singularityware/singularity/docs"
+	// "github.com/singularityware/singularity/docs"
 )
 
 var pullUse string = `pull [pull options...] [library://[user[collection/[<container>:tag]]]]`
@@ -44,12 +44,7 @@ var (
 )
 
 func init() {
-	manHelp := func(c *cobra.Command, args []string) {
-		docs.DispManPg("singularity-pull")
-	}
-
 	pullCmd.Flags().SetInterspersed(false)
-	pullCmd.SetHelpFunc(manHelp)
 	SingularityCmd.AddCommand(pullCmd)
 
 	pullCmd.Flags().BoolVarP(&Force, "force", "F", false, "overwrite an image file if it exists")
